requirements.txt
simple-salesforce
cryptography
pyjwt
requests
python-dotenv
paramiko